reworked procedural gen to work with preset rooms instead

+++ Assets/FishNet/Runtime/Managing/Timing/TimeManager.cs
@@ -910,8 +910,8 @@ namespace FishNet.Managing.Timing
         public double TimePassed(PreciseTick preciseTick, bool allowNegative = false)
         {
             PreciseTick currentPt = GetPreciseTick(TickType.Tick);
-
-            long tickDifference = (currentPt.Tick - preciseTick.Tick);
+            
+            long tickDifference = ((long)currentPt.Tick - (long)preciseTick.Tick);
             double percentDifference = (currentPt.PercentAsDouble - preciseTick.PercentAsDouble);
 
             /* If tickDifference is less than 0 or tickDifference and percentDifference are 0 or less
